move pubnub to setup folder
